+++ src/main/java/api/RequestApi.java
@@ -3,7 +3,7 @@ package api;
 import io.restassured.RestAssured;
 
 public class RequestApi {
-    public void setUp() {
-        RestAssured.baseURI = "https://stellarburgers.nomoreparties.site/api/auth";
+    public void apiEndPoint() {
+        RestAssured.baseURI = "https://stellarburgers.nomoreparties.site/api";
     }
 }
