+++ src/main/java/api/user/UserLoginApi.java
@@ -1,17 +1,18 @@
-package api;
+package api.user;
 
+import api.RequestApi;
 import io.restassured.response.Response;
 
 import static io.restassured.RestAssured.given;
 
-public class UserLoginApi extends RequestApi{
+public class UserLoginApi extends RequestApi {
     public String requestUserLogin(Object model, Integer code) {
         Response response = given()
                 .header("Content-type", "application/json")
                 .and()
                 .body(model)
                 .when()
-                .post("/login");
+                .post("/auth/login");
         response.then().statusCode(code);
         return response.getBody().asString();
     }
