+++ src/main/java/functions/orders/OrdersCreateFunctions.java
@@ -1,0 +1,6 @@
+package functions.orders;
+
+import api.orders.OrdersCreateApi;
+
+public class OrdersCreateFunctions extends OrdersCreateApi {
+}
