+++ src/main/java/functions/user/UserLoginFunctions.java
@@ -1,7 +1,7 @@
-package functions;
+package functions.user;
 
-import api.UserLoginApi;
-import models.request.UserLoginRequestModel;
+import api.user.UserLoginApi;
+import models.request.user.UserLoginRequestModel;
 
 public class UserLoginFunctions extends UserLoginApi {
     public String getUserLogin(String email, String password, Integer code){
