+++ src/main/java/functions/user/UserUpdateFunctions.java
@@ -1,7 +1,7 @@
-package functions;
+package functions.user;
 
-import api.UserUpdateApi;
-import models.request.UserUpdateRequestModel;
+import api.user.UserUpdateApi;
+import models.request.user.UserUpdateRequestModel;
 
 public class UserUpdateFunctions extends UserUpdateApi {
     public String getUserUpdate(String name, String email, String token, Integer code){
