+++ src/main/java/models/request/user/UserCreateRequestModel.java
@@ -1,4 +1,4 @@
-package models.request;
+package models.request.user;
 
 public class UserCreateRequestModel {
     public String name;
