+++ src/main/java/models/request/user/UserUpdateRequestModel.java
@@ -1,4 +1,4 @@
-package models.request;
+package models.request.user;
 
 public class UserUpdateRequestModel {
     public String name;
