+++ src/main/java/models/response/user/UserModel.java
@@ -1,4 +1,4 @@
-package models.response;
+package models.response.user;
 
 public class UserModel {
     public String name;
