+++ src/main/java/models/response/user/UserUpdateResponseModel.java
@@ -1,4 +1,4 @@
-package models.response;
+package models.response.user;
 
 public class UserUpdateResponseModel {
     public boolean success;
